agentic-ai-finpilot-main/backend/requirements.txt
flask
flask-cors
gunicorn
google-genai
python-dotenv
requests
langchain
langgraph
